requirements.txt
numpy
oauthlib
pandas
requests
requests-oauthlib
urllib3
quickchart.io
disnakes